requirements.txt
pandas
numpy
setuptools
kagglehub
torch
opencv-python-headless
torchvision
tensorboard
dvc
fastapi[all]
uvicorn
pillow